Revert renaming module
