+++ tests/tracing/test_fluent.py
@@ -105,7 +105,7 @@ def test_trace_ignore_exception_from_tracing_logic(mock_client):
     model = TestModel()
 
     # Exception during span creation: no-op span wrapper created and no trace is logged
-    with mock.patch("mlflow.tracing.python_api.get_tracer", side_effect=ValueError("Some error")):
+    with mock.patch("mlflow.tracing.fluent.get_tracer", side_effect=ValueError("Some error")):
         output = model.predict(2, 5)
 
     assert output == 7
